Extract a file per component.
Use for top level of advanced docs.
Need to decide whether we have a common model or just share UI
components.
The second level is a good bit more complicated so I think it'll need
its own component (probably the third too).
