+++ src/documentation/AdvancedDocumentation.tsx
@@ -2,30 +2,42 @@ import { Box } from "@chakra-ui/layout";
 import { Spinner } from "@chakra-ui/spinner";
 import React, { useEffect, useState } from "react";
 import { useIntl } from "react-intl";
-import { apiDocs, ApiDocsResponse } from "../language-server/apidocs";
+import { apiDocs } from "../language-server/apidocs";
 import { useLanguageServerClient } from "../language-server/language-server-hooks";
 import { pullModulesToTop } from "./apidocs-util";
-import ApiDocsAccordion from "./ApiDocsAccordion";
+import { Toolkit } from "./ToolkitDocumentation/model";
+import ToolkitDocumentation from "./ToolkitDocumentation";
 
 const AdvancedDocumentation = () => {
   const client = useLanguageServerClient();
   const intl = useIntl();
-  const [apidocs, setApiDocs] = useState<ApiDocsResponse | undefined>();
+  const [apidocs, setApiDocs] = useState<Toolkit | undefined>();
   useEffect(() => {
     const load = async () => {
       if (client) {
         await client.initialize();
         const docs = await apiDocs(client);
         pullModulesToTop(docs);
-        setApiDocs(docs);
+        const contents = Object.values(docs).map((module) => {
+          return {
+            name: module.fullName,
+            description: module.docString || "",
+            contents: [],
+          };
+        });
+        setApiDocs({
+          name: "Advanced",
+          description: "Reference documentation for MicroPython",
+          contents,
+        });
       }
     };
     load();
   }, [client]);
   return (
-    <Box height="100%" p={2}>
+    <Box>
       {apidocs ? (
-        <ApiDocsAccordion docs={apidocs} />
+        <ToolkitDocumentation toolkit={apidocs} />
       ) : (
         <Spinner
           display="block"
