+++ src/documentation/ToolkitDocumentation/index.ts
@@ -1,0 +1,3 @@
+import { ToolkitDocumentation } from "./ToolkitDocumentation";
+
+export default ToolkitDocumentation;
